moved definitions logic from parser class to utils
